+++ app/src/main/java/org/ole/planet/myplanet/lite/DashboardTeamSurveysFragment.kt
@@ -14,17 +14,10 @@ import android.widget.TextView
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 import androidx.viewpager2.widget.ViewPager2
-import com.google.android.material.badge.BadgeDrawable
 import com.google.android.material.tabs.TabLayout
 import com.google.android.material.tabs.TabLayoutMediator
-import io.noties.markwon.Markwon
-import java.text.DateFormat
-import java.util.Date
-import java.util.Locale
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
@@ -365,349 +358,4 @@ class DashboardTeamSurveysFragment : Fragment(R.layout.fragment_dashboard_team_s
             }
         }
     }
-}
-
-private class SurveysPagerAdapter(
-    private val teamEmptyMessage: String,
-    private val adoptedEmptyMessage: String,
-    private val outboxEmptyMessage: String,
-    private val statusStore: DashboardSurveyStatusStore,
-    private val onSurveySelected: (SurveyDocument) -> Unit,
-    private val onSurveyDownloadRequested: (SurveyDocument) -> Unit,
-    private val onOutboxSelected: (OutboxEntry) -> Unit,
-    private val onStatusChanged: () -> Unit,
-) : RecyclerView.Adapter<SurveysPagerAdapter.PageViewHolder>() {
-
-    private val pages = listOf(Page.TEAM, Page.ADOPTED, Page.OUTBOX)
-    private var teamItems: List<SurveyItemUiModel> = emptyList()
-    private var adoptedItems: List<SurveyItemUiModel> = emptyList()
-    private var outboxItems: List<OutboxEntry> = emptyList()
-
-    override fun onCreateViewHolder(parent: android.view.ViewGroup, viewType: Int): PageViewHolder {
-        val view = android.view.LayoutInflater.from(parent.context)
-            .inflate(R.layout.item_dashboard_surveys_page, parent, false)
-        return PageViewHolder(view)
-    }
-
-    override fun getItemCount(): Int = pages.size
-
-    override fun onBindViewHolder(holder: PageViewHolder, position: Int) {
-        val page = pages[position]
-        when (page) {
-            Page.TEAM -> holder.bindTeam(
-                items = teamItems,
-                emptyMessage = teamEmptyMessage,
-                statusStore = statusStore,
-                onStatusChanged = onStatusChanged,
-                onSurveySelected = onSurveySelected,
-                onDownloadRequested = onSurveyDownloadRequested,
-            )
-            Page.ADOPTED -> holder.bindTeam(
-                items = adoptedItems,
-                emptyMessage = adoptedEmptyMessage,
-                statusStore = statusStore,
-                onStatusChanged = onStatusChanged,
-                onSurveySelected = onSurveySelected,
-                onDownloadRequested = onSurveyDownloadRequested,
-            )
-            Page.OUTBOX -> holder.bindOutbox(
-                items = outboxItems,
-                emptyMessage = outboxEmptyMessage,
-                onOutboxSelected = onOutboxSelected,
-            )
-        }
-    }
-
-    fun submit(
-        teamSurveys: List<SurveyDocument>,
-        adoptedSurveys: List<SurveyDocument>,
-        completionCounts: Map<String, Int>,
-        savedSurveys: Set<String>,
-        savedRevisions: Map<String, String?>,
-        outboxItems: List<OutboxEntry>,
-    ) {
-        teamItems = teamSurveys.map { it.toUiModel(completionCounts, savedSurveys, savedRevisions) }
-        adoptedItems = adoptedSurveys.map { it.toUiModel(completionCounts, savedSurveys, savedRevisions) }
-        this.outboxItems = outboxItems
-        notifyItemChanged(0)
-        notifyItemChanged(1)
-        notifyItemChanged(2)
-    }
-
-    fun updateSavedSurveys(savedSurveys: Set<String>, savedRevisions: Map<String, String?>) {
-        teamItems = teamItems.map { it.copy(isSaved = savedSurveys.contains(it.document.id.orEmpty()), savedRevision = savedRevisions[it.document.id.orEmpty()]) }
-        adoptedItems = adoptedItems.map { it.copy(isSaved = savedSurveys.contains(it.document.id.orEmpty()), savedRevision = savedRevisions[it.document.id.orEmpty()]) }
-        notifyItemChanged(0)
-        notifyItemChanged(1)
-    }
-
-    private fun SurveyDocument.toUiModel(
-        completionCounts: Map<String, Int>,
-        savedSurveys: Set<String>,
-        savedRevisions: Map<String, String?>
-    ): SurveyItemUiModel {
-        val documentId = id.orEmpty()
-        return SurveyItemUiModel(
-            document = this,
-            completionCount = completionCounts[documentId] ?: 0,
-            isSaved = savedSurveys.contains(documentId),
-            savedRevision = savedRevisions[documentId],
-        )
-    }
-
-    private enum class Page { TEAM, ADOPTED, OUTBOX }
-
-    class PageViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-        private val emptyView: TextView = itemView.findViewById(R.id.dashboardSurveysListEmpty)
-        private val recyclerView: RecyclerView = itemView.findViewById(R.id.dashboardSurveysList)
-        private var teamAdapter: SurveyItemsAdapter? = null
-        private var outboxAdapter: SurveyOutboxAdapter? = null
-        private var currentPage: Page = Page.TEAM
-        init {
-            recyclerView.layoutManager = LinearLayoutManager(itemView.context)
-        }
-
-        fun bindTeam(
-            items: List<SurveyItemUiModel>,
-            emptyMessage: String,
-            statusStore: DashboardSurveyStatusStore,
-            onStatusChanged: () -> Unit,
-            onSurveySelected: (SurveyDocument) -> Unit,
-            onDownloadRequested: (SurveyDocument) -> Unit,
-        ) {
-            if (currentPage != Page.TEAM && currentPage != Page.ADOPTED) {
-                recyclerView.adapter = null
-            }
-            if (recyclerView.adapter !is SurveyItemsAdapter) {
-                teamAdapter = SurveyItemsAdapter(
-                    statusStore = statusStore,
-                    onStatusChanged = onStatusChanged,
-                    onSurveySelected = onSurveySelected,
-                    onDownloadRequested = onDownloadRequested,
-                )
-                recyclerView.adapter = teamAdapter
-            }
-            teamAdapter?.submitList(items)
-            emptyView.text = emptyMessage
-            emptyView.isVisible = items.isEmpty()
-            recyclerView.isVisible = items.isNotEmpty()
-            outboxAdapter = null
-            currentPage = Page.TEAM
-        }
-
-        fun bindOutbox(
-            items: List<OutboxEntry>,
-            emptyMessage: String,
-            onOutboxSelected: (OutboxEntry) -> Unit,
-        ) {
-            if (currentPage != Page.OUTBOX) {
-                recyclerView.adapter = null
-            }
-            if (recyclerView.adapter !is SurveyOutboxAdapter) {
-                outboxAdapter = SurveyOutboxAdapter(onOutboxSelected)
-                recyclerView.adapter = outboxAdapter
-            }
-            outboxAdapter?.submitList(items)
-            emptyView.text = emptyMessage
-            emptyView.isVisible = items.isEmpty()
-            recyclerView.isVisible = items.isNotEmpty()
-            teamAdapter = null
-            currentPage = Page.OUTBOX
-        }
-    }
-}
-
-private data class SurveyItemUiModel(
-    val document: SurveyDocument,
-    val completionCount: Int,
-    val isSaved: Boolean,
-    val savedRevision: String?,
-)
-
-private class SurveyItemUiModelDiffCallback : androidx.recyclerview.widget.DiffUtil.ItemCallback<SurveyItemUiModel>() {
-    override fun areItemsTheSame(oldItem: SurveyItemUiModel, newItem: SurveyItemUiModel): Boolean = oldItem.document.id == newItem.document.id
-    override fun areContentsTheSame(oldItem: SurveyItemUiModel, newItem: SurveyItemUiModel): Boolean = oldItem == newItem
-}
-
-private class SurveyItemsAdapter(
-    private val statusStore: DashboardSurveyStatusStore,
-    private val onStatusChanged: () -> Unit,
-    private val onSurveySelected: (SurveyDocument) -> Unit,
-    private val onDownloadRequested: (SurveyDocument) -> Unit,
-) : androidx.recyclerview.widget.ListAdapter<SurveyItemUiModel, SurveyItemsAdapter.SurveyViewHolder>(SurveyItemUiModelDiffCallback()) {
-
-    override fun onCreateViewHolder(parent: android.view.ViewGroup, viewType: Int): SurveyViewHolder {
-        val view = android.view.LayoutInflater.from(parent.context)
-            .inflate(R.layout.item_dashboard_survey, parent, false)
-        return SurveyViewHolder(
-            view,
-            statusStore,
-            onStatusChanged,
-            onSurveySelected,
-            onDownloadRequested,
-        )
-    }
-
-    override fun onBindViewHolder(holder: SurveyViewHolder, position: Int) {
-        holder.bind(getItem(position))
-    }
-
-    class SurveyViewHolder(
-        itemView: View,
-        private val statusStore: DashboardSurveyStatusStore,
-        private val onStatusChanged: () -> Unit,
-        private val onSurveySelected: (SurveyDocument) -> Unit,
-        private val onDownloadRequested: (SurveyDocument) -> Unit,
-    ) : RecyclerView.ViewHolder(itemView) {
-        private val title: TextView = itemView.findViewById(R.id.dashboardSurveyTitle)
-        private val description: TextView = itemView.findViewById(R.id.dashboardSurveyDescription)
-        private val statusIcon: android.widget.ImageView = itemView.findViewById(R.id.dashboardSurveyStatusIcon)
-        private val metadata: TextView = itemView.findViewById(R.id.dashboardSurveyMetadata)
-        private val markwon: Markwon = Markwon.builder(itemView.context).build()
-        private val downloadButton: com.google.android.material.button.MaterialButton =
-            itemView.findViewById(R.id.dashboardSurveyDownloadButton)
-
-        fun bind(uiModel: SurveyItemUiModel) {
-            val document = uiModel.document
-            markwon.setMarkdown(title, document.name.orEmpty().replace("\n", "  \n"))
-            val detail = document.description?.takeIf { it.isNotBlank() }
-            if (detail.isNullOrBlank()) {
-                description.isVisible = false
-                description.text = ""
-            } else {
-                description.isVisible = true
-                markwon.setMarkdown(description, detail.replace("\n", "  \n"))
-            }
-
-            val created = formatCreatedDate(document.createdDate)
-            val completions = uiModel.completionCount
-            metadata.text = itemView.context.getString(
-                R.string.dashboard_surveys_metadata,
-                created,
-                completions,
-            )
-            metadata.isVisible = true
-
-            val isSaved = uiModel.isSaved
-            val savedRevision = uiModel.savedRevision
-            val latestRevision = document.rev
-            val isOutdated = isSaved && !latestRevision.isNullOrBlank() && savedRevision != null && savedRevision != latestRevision
-            val context = itemView.context
-            val buttonEnabled = !isSaved || isOutdated
-            val iconRes = if (isOutdated) R.drawable.ic_sync_again_24 else R.drawable.ic_survey_download_24
-            downloadButton.setIconResource(iconRes)
-            downloadButton.isEnabled = buttonEnabled
-            downloadButton.alpha = if (buttonEnabled) 1f else 0.7f
-            downloadButton.contentDescription = context.getString(
-                if (buttonEnabled) R.string.dashboard_survey_download else R.string.dashboard_survey_downloaded,
-            )
-            downloadButton.setOnClickListener {
-                if (buttonEnabled) {
-                    onDownloadRequested(document)
-                }
-            }
-
-            var status = statusStore.getStatus(document.id) ?: SurveyStatus.NEW
-            statusIcon.setImageResource(status.iconRes)
-            val openSurvey = {
-                if (status != SurveyStatus.COMPLETED) {
-                    statusStore.markViewed(document.id)
-                    status = SurveyStatus.VIEWED
-                    statusIcon.setImageResource(status.iconRes)
-                    onStatusChanged()
-                }
-                onSurveySelected(document)
-            }
-            itemView.setOnClickListener { openSurvey() }
-            title.setOnClickListener { openSurvey() }
-            description.setOnClickListener { openSurvey() }
-            itemView.setOnLongClickListener {
-                statusStore.markCompleted(document.id)
-                status = SurveyStatus.COMPLETED
-                statusIcon.setImageResource(status.iconRes)
-                onStatusChanged()
-                true
-            }
-        }
-
-        private fun formatCreatedDate(raw: String?): String {
-            val context = itemView.context
-            val millis = raw?.toLongOrNull() ?: return context.getString(
-                R.string.dashboard_surveys_unknown_date,
-            )
-            return runCatching {
-                DateFormat.getDateInstance(DateFormat.MEDIUM, Locale.getDefault())
-                    .format(Date(millis))
-            }.getOrElse {
-                context.getString(R.string.dashboard_surveys_unknown_date)
-            }
-        }
-    }
-}
-
-private val SurveyStatus.iconRes: Int
-    get() = when (this) {
-        SurveyStatus.NEW -> R.drawable.ic_survey_new_24
-        SurveyStatus.VIEWED -> R.drawable.ic_survey_viewed_24
-        SurveyStatus.COMPLETED -> R.drawable.ic_survey_completed_24
-    }
-
-private class OutboxEntryDiffCallback : androidx.recyclerview.widget.DiffUtil.ItemCallback<OutboxEntry>() {
-    override fun areItemsTheSame(oldItem: OutboxEntry, newItem: OutboxEntry): Boolean = oldItem.id == newItem.id
-    override fun areContentsTheSame(oldItem: OutboxEntry, newItem: OutboxEntry): Boolean = oldItem == newItem
-}
-
-private class SurveyOutboxAdapter(
-    private val onOutboxSelected: (OutboxEntry) -> Unit,
-) : androidx.recyclerview.widget.ListAdapter<OutboxEntry, SurveyOutboxAdapter.OutboxViewHolder>(OutboxEntryDiffCallback()) {
-
-    override fun onCreateViewHolder(parent: android.view.ViewGroup, viewType: Int): OutboxViewHolder {
-        val view = android.view.LayoutInflater.from(parent.context)
-            .inflate(R.layout.item_dashboard_survey_outbox, parent, false)
-        return OutboxViewHolder(view, onOutboxSelected)
-    }
-
-    override fun onBindViewHolder(holder: OutboxViewHolder, position: Int) {
-        holder.bind(getItem(position))
-    }
-
-
-    class OutboxViewHolder(
-        itemView: View,
-        private val onOutboxSelected: (OutboxEntry) -> Unit,
-    ) : RecyclerView.ViewHolder(itemView) {
-        private val title: TextView = itemView.findViewById(R.id.dashboardOutboxTitle)
-        private val subtitle: TextView = itemView.findViewById(R.id.dashboardOutboxSubtitle)
-        private val timestamp: TextView = itemView.findViewById(R.id.dashboardOutboxTimestamp)
-
-        fun bind(entry: OutboxEntry) {
-            title.text = entry.surveyName.orEmpty().ifBlank { itemView.context.getString(R.string.dashboard_outbox_unknown_survey) }
-            subtitle.text = entry.teamName.orEmpty().ifBlank { itemView.context.getString(R.string.dashboard_outbox_unknown_team) }
-            val formattedDate = runCatching {
-                DateFormat.getDateTimeInstance(
-                    DateFormat.MEDIUM,
-                    DateFormat.SHORT,
-                    Locale.getDefault(),
-                ).format(Date(entry.createdAt))
-            }.getOrElse {
-                ""
-            }
-            timestamp.text = itemView.context.getString(R.string.dashboard_outbox_saved_at, formattedDate)
-            timestamp.isVisible = formattedDate.isNotBlank()
-            itemView.setOnClickListener { onOutboxSelected(entry) }
-        }
-    }
-}
-
-private fun TabLayout.Tab.ensureOffsetBadge(count: Int) {
-    val badge = orCreateBadge
-    badge.number = count
-    badge.badgeGravity = BadgeDrawable.TOP_END
-    val resources = view.resources
-    badge.horizontalOffset = resources.getDimensionPixelOffset(
-        R.dimen.dashboard_surveys_tab_badge_horizontal_offset,
-    )
-    badge.verticalOffset = resources.getDimensionPixelOffset(
-        R.dimen.dashboard_surveys_tab_badge_vertical_offset,
-    )
 }
